requirements.txt
asarPy==1.0.1
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
SQLAlchemy==2.0.34
typing_extensions==4.12.2
Werkzeug==3.0.4
WTForms==3.1.2
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Pillow==8.2.0
email_validator==1.1.2
